bring publish into correspondance with ship by moving failure into publish_pattern

+++ bin/pattern/publish_pattern.sh
@@ -1,6 +1,11 @@
 #!/usr/bin/env sh
 
 publish_pattern() {
+	if [[ ${PATTERN} == "" ]] ; then
+		printf "Please specify a pattern to publish with 'PATTERN=my-pattern'"
+		exit 1
+	fi
+
 	cp ../../bin/pattern/tsconfig.json .
 
 	rm -r package/dist/* || true
